fix: record_wager odds type + Opportunity.potential_revenue default
- record_wager: pass int(odds) instead of float(odds) to match
  PlacedWager.odds IntegerField
- Opportunity.potential_revenue: add default=0 to prevent NOT NULL
  violations on auto-created records
- content_review_tool: already handles approve/reject aliases
  (Session 1075) — no changes needed

+++ core/services/tool_dispatcher.py
@@ -7070,7 +7070,7 @@ class ToolDispatcher:
             wager_kwargs = {
                 'notes': description,
                 'stake': stake,
-                'odds': float(odds),
+                'odds': int(odds),  # PlacedWager.odds is IntegerField
                 'potential_payout': potential_payout,
                 'status': 'pending',
             }
